requirements.txt
accelerate==1.1.1
langchain-ollama==0.3.3
numpy==1.26.4
pandas==2.2.3
Pillow==12.0.0
PyYAML==6.0.2
tqdm==4.67.1
transformers==4.51.3
